# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/fees": {
      "get": {
        "operationId": "FeesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "UNPAID",
                "PARTIAL",
                "PAID",
                "OUTSTANDING"
              ]
            }
          },
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "traineeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodStartFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodStartTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedFeeRow"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the fees of the club. Filtered and paginated.",
        "tags": [
          "Fees"
        ]
      },
      "post": {
        "operationId": "FeesController_create",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fee"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create one fee for one trainee.",
        "tags": [
          "Fees"
        ]
      }
    },
    "/api/fees/unbilled": {
      "get": {
        "operationId": "FeesController_unbilled",
        "parameters": [
          {
            "name": "periodStart",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodEnd",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnbilledFeeList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the enrolled trainees that have no fee yet for a month.",
        "tags": [
          "Fees"
        ]
      }
    },
    "/api/fees/{id}": {
      "get": {
        "operationId": "FeesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read one fee with its payments and its refunds.",
        "tags": [
          "Fees"
        ]
      },
      "patch": {
        "operationId": "FeesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fee"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change one fee. The amount cannot go below what is already paid.",
        "tags": [
          "Fees"
        ]
      },
      "delete": {
        "operationId": "FeesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete one fee. Its payments and refunds go with it.",
        "tags": [
          "Fees"
        ]
      }
    },
    "/api/fees/generate-monthly": {
      "post": {
        "operationId": "FeesController_generateMonthly",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateMonthlyFeesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Make the monthly fees of a class for one month, in one request.",
        "tags": [
          "Fees"
        ]
      }
    },
    "/api/fees/generate-session": {
      "post": {
        "operationId": "FeesController_generateSession",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSessionFeesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Make the per-session fees for the sessions of a period.",
        "tags": [
          "Fees"
        ]
      }
    },
    "/api/fees/generate-course": {
      "post": {
        "operationId": "FeesController_generateCourse",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCourseFeesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Make one course fee for each enrolled trainee of a class.",
        "tags": [
          "Fees"
        ]
      }
    },
...
  },
  "info": {
    "title": "PulseDesk API",
...
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateFeeDto": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "string"
          },
          "traineeId": {
            "type": "string"
          },
          "periodStart": {
            "type": "string"
          },
          "periodEnd": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01,
            "maximum": 1000000
          },
          "notes": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "classId",
          "traineeId",
          "periodStart",
          "periodEnd",
          "amount"
        ]
      },
      "UpdateFeeDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0.01,
            "maximum": 1000000
          },
          "periodStart": {
            "type": "string"
          },
          "periodEnd": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
      "GenerateMonthlyFeesDto": {
        "type": "object",
        "properties": {
          "periodStart": {
            "type": "string"
          },
          "periodEnd": {
            "type": "string"
          },
          "classId": {
            "type": "string"
          }
        },
        "required": [
          "periodStart",
          "periodEnd"
        ]
      },
      "GenerateSessionFeesDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "classId": {
            "type": "string"
          }
        },
        "required": [
          "from",
          "to"
        ]
      },
      "GenerateCourseFeesDto": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "string"
          }
        }
      },
...
      "GenerateResult": {
        "type": "object",
        "properties": {
          "created": {
            "type": "number"
          },
          "skipped": {
            "type": "number"
          }
        },
        "required": [
          "created",
          "skipped"
        ],
        "additionalProperties": false
      },
...
      "PaginatedFeeRow": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "tenantId": {
                  "type": "string"
                },
                "classId": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "traineeId": {
                  "type": "string"
                },
                "sessionId": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "periodStart": {
                  "type": "string"
                },
                "periodEnd": {
                  "type": "string"
                },
                "amount": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "UNPAID",
                    "PARTIAL",
                    "PAID"
                  ]
                },
                "notes": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                },
                "paid": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "tenantId",
                "classId",
                "traineeId",
                "sessionId",
                "periodStart",
                "periodEnd",
                "amount",
                "status",
                "notes",
                "createdAt",
                "updatedAt",
                "paid"
              ],
              "additionalProperties": false
            }
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "additionalProperties": false
      },
      "Fee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "classId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "traineeId": {
            "type": "string"
          },
          "sessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "periodStart": {
            "type": "string"
          },
          "periodEnd": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "UNPAID",
              "PARTIAL",
              "PAID"
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "classId",
          "traineeId",
          "sessionId",
          "periodStart",
          "periodEnd",
          "amount",
          "status",
          "notes",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "UnbilledFeeList": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "classId": {
              "type": "string"
            },
            "className": {
              "type": "string"
            },
            "traineeId": {
              "type": "string"
            },
            "traineeFirstName": {
              "type": "string"
            },
            "traineeLastName": {
              "type": "string"
            },
            "amount": {
              "type": "string"
            }
          },
          "required": [
            "classId",
            "className",
            "traineeId",
            "traineeFirstName",
            "traineeLastName",
            "amount"
          ],
          "additionalProperties": false
        }
      },
      "FeeDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "classId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "traineeId": {
            "type": "string"
          },
          "sessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "periodStart": {
            "type": "string"
          },
          "periodEnd": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "UNPAID",
              "PARTIAL",
              "PAID"
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "class": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "billingMode": {
                    "type": "string",
                    "enum": [
                      "PER_MONTH",
                      "PER_SESSION",
                      "PER_COURSE"
                    ]
                  }
                },
                "required": [
                  "id",
                  "name",
                  "billingMode"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "trainee": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "firstName",
              "lastName"
            ],
            "additionalProperties": false
          },
          "payments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "tenantId": {
                  "type": "string"
                },
                "feeId": {
                  "type": "string"
                },
                "amount": {
                  "type": "string"
                },
                "paidAt": {
                  "type": "string"
                },
                "method": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "notes": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "recordedById": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "recordedByEmailSnapshot": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "recordedByNameSnapshot": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "createdAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "tenantId",
                "feeId",
                "amount",
                "paidAt",
                "method",
                "notes",
                "recordedById",
                "recordedByEmailSnapshot",
                "recordedByNameSnapshot",
                "createdAt"
              ],
              "additionalProperties": false
            }
          },
          "refunds": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "tenantId": {
                  "type": "string"
                },
                "feeId": {
                  "type": "string"
                },
                "amount": {
                  "type": "string"
                },
                "refundedAt": {
                  "type": "string"
                },
                "method": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "notes": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "recordedById": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "recordedByEmailSnapshot": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "recordedByNameSnapshot": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "createdAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "tenantId",
                "feeId",
                "amount",
                "refundedAt",
                "method",
                "notes",
                "recordedById",
                "recordedByEmailSnapshot",
                "recordedByNameSnapshot",
                "createdAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "tenantId",
          "classId",
          "traineeId",
          "sessionId",
          "periodStart",
          "periodEnd",
          "amount",
          "status",
          "notes",
          "createdAt",
          "updatedAt",
          "class",
          "trainee",
          "payments",
          "refunds"
        ],
        "additionalProperties": false
      },
...
    }
  }
}